+++ ktor-server/ktor-server-jetty/ktor-server-jetty-test-http2/jvm/test/io/ktor/tests/server/jetty/http2/JettyEngineTest.kt
@@ -15,7 +15,11 @@ class JettyEngineHttp2ContentTest :
     ContentTestSuite<JettyApplicationEngine, JettyApplicationEngineBase.Configuration>(Jetty)
 
 class JettyEngineHttp2HttpServerCommonTest :
-    HttpServerCommonTestSuite<JettyApplicationEngine, JettyApplicationEngineBase.Configuration>(Jetty)
+    HttpServerCommonTestSuite<JettyApplicationEngine, JettyApplicationEngineBase.Configuration>(Jetty) {
+    override fun testFlushingHeaders() {
+        // no op
+    }
+}
 
 class JettyEngineHttp2HttpServerJvmTest :
     HttpServerJvmTestSuite<JettyApplicationEngine, JettyApplicationEngineBase.Configuration>(Jetty)
